FastExamSystem/pyproject.toml
[project]
name = "fastexamsystem"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-mistralai>=0.2.10",
    "psycopg2>=2.9.10",
    "pwdlib[argon2,bcrypt]>=0.2.1",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
]
